# api/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "AI-Generated Image Detector API",
        "description": "Production inference-only API for estimating synthetic image probability",
        "version": "1.0.0"
    },
    "paths": {
        "/analyze/image": {
            "post": {
                "summary": "Analyze Image",
                "description": "Analyze a single image for synthetic probability. Returns probability estimate with uncertainty and limitations. Does NOT return binary verdicts.",
                "operationId": "analyze_image",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "required": [
                                    "file"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Analysis result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnalysisResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Server error"
                    },
                    "503": {
                        "description": "Model not loaded"
                    }
                }
            }
        },
        "/health": {
            "get": {
                "summary": "Health Check",
                "description": "Service health check. No model inference performed.",
                "operationId": "health_check",
                "responses": {
                    "200": {
                        "description": "Health status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HealthResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/capabilities": {
            "get": {
                "summary": "Get Capabilities",
                "description": "Describe system capabilities and limitations.",
                "operationId": "get_capabilities",
                "responses": {
                    "200": {
                        "description": "Capabilities description",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CapabilitiesResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AnalysisResponse": {
                "type": "object",
                "required": [
                    "synthetic_probability",
                    "confidence_interval",
                    "interpretation",
                    "uncertainty_notice",
                    "model_version",
                    "semantic_version",
                    "api_version",
                    "request_id",
                    "timestamp"
                ],
                "properties": {
                    "synthetic_probability": {
                        "type": "number",
                        "minimum": 0,
                        "maximum": 1
                    },
                    "confidence_interval": {
                        "type": "object",
                        "properties": {
                            "lower": {
                                "type": "number"
                            },
                            "upper": {
                                "type": "number"
                            },
                            "level": {
                                "type": "number"
                            }
                        }
                    },
                    "interpretation": {
                        "type": "string"
                    },
                    "uncertainty_notice": {
                        "type": "string"
                    },
                    "model_version": {
                        "type": "string"
                    },
                    "semantic_version": {
                        "type": "string"
                    },
                    "api_version": {
                        "type": "string"
                    },
                    "request_id": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string"
                    }
                }
            },
            "HealthResponse": {
                "type": "object",
                "properties": {
                    "service_status": {
                        "type": "string"
                    },
                    "model_loaded": {
                        "type": "boolean"
                    },
                    "uptime_seconds": {
                        "type": "number"
                    },
                    "model_version": {
                        "type": "string"
                    },
                    "semantic_version": {
                        "type": "string"
                    },
                    "api_version": {
                        "type": "string"
                    }
                }
            },
            "CapabilitiesResponse": {
                "type": "object",
                "properties": {
                    "can_do": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "cannot_do": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "known_limitations": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "intended_use": {
                        "type": "string"
                    },
                    "model_version": {
                        "type": "string"
                    },
                    "semantic_version": {
                        "type": "string"
                    },
                    "api_version": {
                        "type": "string"
                    }
                }
            },
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "request_id": {
                        "type": "string"
                    }
                }
            }
        }
    }
}